core: remove interdependencies between modules
