+++ core/Bool.carp
@@ -3,11 +3,11 @@
 (defmodule Bool
   (register = (Fn [Bool Bool] Bool))
   (register /= (Fn [Bool Bool] Bool))
-  (register str (Fn [Bool] String))
   (register copy (Fn [&Bool] Bool))
   (register format (Fn [&String Bool] String))
-  )
+)
 
 (defmodule RefBool
   (defn = [a b]
     (Bool.= @a @b)))
+)
