+++ core/Double.carp
@@ -15,8 +15,6 @@
   (register from-int (Fn [Int] Double))
   (register to-float (Fn [Double] Float))
   (register from-float (Fn [Float] Double))
-  (register random (Fn [] Double))
-  (register random-between (Fn [Double Double] Double))
   (register sin (Fn [Double] Double))
   (register cos (Fn [Double] Double))
   (register tan (Fn [Double] Double))
@@ -35,13 +33,11 @@
   (register modf (Fn [Double (Ref Double)] Double))
   (register pow (Fn [Double Double] Double))
   (register sqrt (Fn [Double] Double))
-  (register str (Fn [Double] String))
   (register floor (Fn [Double] Double))
   (register ceil (Fn [Double] Double))
   (register mod (Fn [Double Double] Double))
   (register copy (Fn [(Ref Double)] Double))
   (register abs (Fn [Double] Double))
-  (register format (Fn [&String Double] String))
 
   (defn clamp [min, max, val]
     (if (> val max)
