+++ core/Int.carp
@@ -10,11 +10,6 @@
   (register = (λ [Int Int] Bool))
   (register /= (λ [Int Int] Bool))
   (register mod (λ [Int Int] Int))
-  (register seed (λ [Int] ()))
-  (register random (λ [] Int))
-  (register random-between (λ [Int Int] Int))
-  (register str (λ [Int] String))
-  (register from-string (λ [&String] Int))
   (register bit-shift-left (λ [Int Int] Int))
   (register bit-shift-right (λ [Int Int] Int))
   (register bit-and (λ [Int Int] Int))
@@ -24,7 +19,6 @@
   (register inc (λ [Int] Int))
   (register dec (λ [Int] Int))
   (register copy (λ [&Int] Int)) ;; TODO: Should not be needed when refs to value types are auto-converted to non-refs.
-  (register format (Fn [&String Int] String))
 
   (not-on-windows
     (register safe-add (λ [Int Int (Ref Int)] Bool))
