+++ core/Long.carp
@@ -10,10 +10,6 @@
   (register = (λ [Long Long] Bool))
   (register mod (λ [Long Long] Long))
   (register seed (λ [Long] ()))
-  (register random (λ [] Long))
-  (register random-between (λ [Long Long] Long))
-  (register str (λ [Long] String))
-  (register from-string (λ [&String] Long))
   (register bit-shift-left (λ [Long Long] Long))
   (register bit-shift-right (λ [Long Long] Long))
   (register bit-and (λ [Long Long] Long))
@@ -25,7 +21,6 @@
   (register to-int (λ [Long] Int))
   (register from-int (λ [Int] Long))
   (register copy (λ [&Long] Long)) ;; TODO: Should not be needed when refs to value types are auto-converted to non-refs.
-  (register format (Fn [&String Long] String))
 
   (defn /= [x y]
     (not (Long.= x y)))
