+++ core/String.carp
@@ -80,9 +80,6 @@
   (defn empty? [s]
     (Int.= (count s) 0))
 
-  (defn random-sized [n]
-    (from-chars (Array.repeat n Char.random)))
-
   (defn substring [s a b]
     (from-chars (Array.subarray &(chars s) a b)))
 
@@ -125,3 +122,35 @@
 (defmodule Bool (defn prn [x] (Bool.str x)))
 (defmodule Char (defn prn [x] (Char.str x)))
 (defmodule Array (defn prn [x] (Array.str x)))
+
+(defmodule Bool
+  (register str (Fn [Bool] String))
+  (register format (Fn [&String Bool] String))
+)
+
+(defmodule Int
+  (register str (Fn [Int] String))
+  (register format (Fn [&String Int] String))
+  (register from-string (λ [&String] Int))
+)
+
+(defmodule Float
+  (register str (Fn [Float] String))
+  (register format (Fn [&String Float] String))
+)
+
+(defmodule Long
+  (register str (Fn [Long] String))
+  (register format (Fn [&String Long] String))
+  (register from-string (λ [&String] Long))
+)
+
+(defmodule Double
+  (register str (Fn [Double] String))
+  (register format (Fn [&String Double] String))
+)
+
+(defmodule Char
+  (register str (Fn [Char] String))
+  (register format (Fn [&String Char] String))
+)
