+++ examples/guessing.carp
@@ -37,7 +37,7 @@
 
 (defn main []
   (do (println "Seeding random number generator...\n")
-      (seed (System.time))
+      (System.srand (System.time))
       (start-new-game!)
       (while guessing
         (let [user-input (get-line)
